+++ R/addGps.R
@@ -15,6 +15,8 @@
 #' @param thresh maximum time apart in seconds for matching GPS coordinates to
 #'   data, if the closest coordinate is more than \code{thresh} apart then the
 #'   Latitude and Longitude values will be set to \code{NA}
+#' @param keepDiff logical flag to keep time difference column (between GPS time
+#'   and data time)
 #' @param \dots additional arguments for other methods
 #'
 #' @details Latitude and Longitude coordinates will be matched to the data
@@ -58,9 +60,10 @@ setGeneric('addGps', function(x, gps=NULL, thresh = 3600, ...) standardGeneric('
 #' @importFrom dplyr mutate select
 #' @importFrom utils globalVariables
 #' @importFrom stats approxfun
+#' @importFrom geosphere distGeo
 #' @export
 #'
-setMethod('addGps', 'data.frame', function(x, gps, thresh = 3600, ...) {
+setMethod('addGps', 'data.frame', function(x, gps, thresh = 3600, keepDiff=FALSE, ...) {
     # Check for right columns and proper types
     needCols <- c('UTC', 'Latitude', 'Longitude')
     missingCols <- needCols[!(needCols %in% colnames(gps))]
@@ -77,6 +80,7 @@ setMethod('addGps', 'data.frame', function(x, gps, thresh = 3600, ...) {
     # code for just interping:
     # can just make functions once at start and pass along
     gps <- distinct(gps)
+    gpsRange <- range(gps$UTC, na.rm=TRUE)
     interpMethod <- ifelse(nrow(gps) > 1, 'linear', 'constant')
     latFun <- approxfun(x=gps$UTC, y=gps$Latitude, rule=2, method=interpMethod, ties='ordered')
     lonFun <- approxfun(x=gps$UTC, y=gps$Longitude, rule=2, method=interpMethod, ties='ordered')
@@ -93,7 +97,7 @@ setMethod('addGps', 'data.frame', function(x, gps, thresh = 3600, ...) {
     # x[, dataTime := UTC]
     x$dataTime <- x$UTC
     gps <- checkGpsKey(gps)
-
+    
     if('Channel' %in% colnames(gps)) {
         gpsCols <- c('Channel', needCols)
         # gps <- gps[, c('Channel', 'gpsTime', needCols), with=FALSE]
@@ -101,7 +105,7 @@ setMethod('addGps', 'data.frame', function(x, gps, thresh = 3600, ...) {
         gpsCols <- needCols
         # gps <- gps[, c('gpsTime', needCols), with=FALSE]
     }
-
+    
     gps <- gps[, gpsCols, with=FALSE]
     # gps[, gpsTime := UTC]
     gps$gpsTime <- gps$UTC
@@ -114,21 +118,37 @@ setMethod('addGps', 'data.frame', function(x, gps, thresh = 3600, ...) {
         setkeyv(gps, 'UTC') # removing channel key from gps if its there i guess
     }
     result <- gps[x, roll='nearest']
+    result$closeLat <- result$Latitude
+    result$closeLon <- result$Longitude
+    # result[, closeLat := Latitude]
+    # result[, closeLon := Longitude]
     tooFar <- abs(as.numeric(difftime(result$dataTime, result$gpsTime, units='secs'))) > thresh
+    if(keepDiff) {
+        result$timeDiff <- abs(as.numeric(difftime(result$dataTime, result$gpsTime, units='secs')))
+    }
     result[tooFar, c('Latitude', 'Longitude') := NA]
     # result[abs(as.numeric(difftime(dataTime, gpsTime, units='secs'))) > thresh, c('Latitude', 'Longitude') := NA]
     # result$Latitude[tooFar] <- NA
     # result$Longitude[tooFar] <- NA
     result$Latitude[!tooFar] <- latFun(result$dataTime[!tooFar])
     result$Longitude[!tooFar] <- lonFun(result$dataTime[!tooFar])
+    result$gpsUncertainty <- NA
+    result$gpsUncertainty[!tooFar] <- 
+        distGeo(matrix(c(result$Longitude[!tooFar], result$Latitude[!tooFar]), ncol=2),
+                           matrix(c(result$closeLon[!tooFar], result$closeLat[!tooFar]), ncol=2))
     result[, UTC := dataTime]
+    # if were interp'ing out of bounds this ends up at 0 which is incorrect
+    if(interpMethod == 'linear') {
+        result$gpsUncertainty[result$UTC <= gpsRange[1]] <- NA
+        result$gpsUncertainty[result$UTC >= gpsRange[2]] <- NA
+    }
     # result$UTC <- result$dataTime
     if(any(is.na(result$Longitude))) {
-        avgTime <- mean(abs(as.numeric(difftime(result$dataTime[tooFar], result$gpsTime[tooFar], units='secs'))))
-        pamWarning('Some GPS coordinate matches exceeded time threshold, setting',
-                ' value to NA. (Average ', round(avgTime,0), ' seconds apart)')
+        avgTime <- mean(abs(as.numeric(difftime(result$dataTime[tooFar], result$gpsTime[tooFar], units='hours'))))
+        warning('Some GPS coordinate matches exceeded time threshold, setting',
+                ' value to NA. (Average ', round(avgTime,1), ' hours apart)')
     }
-    result[, c('gpsTime', 'dataTime') := NULL]
+    result[, c('gpsTime', 'dataTime', 'closeLat', 'closeLon') := NULL]
     attr(result, 'calltype') <- thisType
     setDF(result)
     result
@@ -149,7 +169,23 @@ setMethod('addGps', signature(x='AcousticEvent'), function(x, gps=NULL, thresh =
         pamWarning('No gps data found for event ', id(x))
         return(x)
     }
-    x@detectors <- addGps(x@detectors, gps, thresh, ...)
+    diffs <- numeric(0)
+    for(d in seq_along(detectors(x))) {
+        withGps <- suppressWarnings(
+            addGps(x@detectors[[d]], gps, thresh, keepDiff=TRUE, ...)
+        )
+        hasNa <- is.na(withGps$Latitude) | is.na(withGps$Longitude)
+        if(any(hasNa)) {
+            diffs <- c(diffs, withGps$timeDiff[hasNa])
+        }
+        withGps$timeDiff <- NULL
+        x@detectors[[d]] <- withGps
+    }
+    if(length(diffs) > 0) {
+        pamWarning(length(diffs), ' GPS matches in event ', id(x), ' exceeded time threshold, setting',
+                   ' to NA. (Average ', round(mean(diffs/3600, na.rm=TRUE),1), ' hours apart)',
+                   which=sys.nframe()-1)
+    }
     x
 })
 
@@ -165,6 +201,7 @@ setMethod('addGps', 'list', function(x, gps, thresh = 3600, ...) {
 #' @export
 #'
 setMethod('addGps', 'AcousticStudy', function(x, gps=NULL, thresh = 3600, ...) {
+    now <- Sys.time()
     if(is.null(gps)) {
         dbExists <- file.exists(files(x)$db)
         if(any(!dbExists)) {
@@ -184,25 +221,43 @@ setMethod('addGps', 'AcousticStudy', function(x, gps=NULL, thresh = 3600, ...) {
             pamWarning('No gps data found in any databases.')
             return(x)
         }
-        gps <- checkGpsKey(gps)
-        # setkeyv(gps, c(key(gps), 'db'))
-        events(x) <- lapply(events(x), function(y) {
+        # gps <- checkGpsKey(gps)
+        # # setkeyv(gps, c(key(gps), 'db'))
+        # events(x) <- lapply(events(x), function(y) {
+        #     thisGps <- gps[gps$db %in% files(y)$db, c('UTC', 'Latitude', 'Longitude'), with=FALSE]
+        #     addGps(y, thisGps, thresh, ...)
+        # })
+        
+    } #else {
+    #     gps <- checkGpsKey(gps)
+    #     events(x) <- lapply(events(x), function(y) {
+    #         if('db' %in% colnames(gps)) {
+    #             thisGps <- gps[gps$db %in% files(y)$db, c('UTC', 'Latitude', 'Longitude'), with=FALSE]
+    #         } else {
+    #             thisGps <- gps
+    #         }
+    #         addGps(y, thisGps, thresh, ...)
+    #     })
+    # }
+    gps <- checkGpsKey(gps)
+    events(x) <- lapply(events(x), function(y) {
+        if('db' %in% colnames(gps)) {
             thisGps <- gps[gps$db %in% files(y)$db, c('UTC', 'Latitude', 'Longitude'), with=FALSE]
-            addGps(y, thisGps, thresh, ...)
-        })
-
-    } else {
-        gps <- checkGpsKey(gps)
-        events(x) <- lapply(events(x), function(y) {
-            if('db' %in% colnames(gps)) {
-                thisGps <- gps[gps$db %in% files(y)$db, c('UTC', 'Latitude', 'Longitude'), with=FALSE]
-            } else {
-                thisGps <- gps
-            }
-            addGps(y, thisGps, thresh, ...)
-            })
-    }
+        } else {
+            thisGps <- gps
+        }
+        suppressWarnings(addGps(y, thisGps, thresh, ...))
+    })
     gps(x) <- gps
+    warns <- mget('PAMWARNING', envir = sys.frame(1), ifnotfound = NA)[['PAMWARNING']]
+    if(is.data.frame(warns)) {
+        warns <- warns[warns$functionName == 'addGps' &
+                           warns$time > now, ]
+        if(nrow(warns) > 0) {
+            pamWarning(nrow(warns), ' events had GPS matches exceeding the time',
+                       ' threshold, use "getWarnings(x)" for details.')
+        }
+    }
     x <- .addPamWarning(x)
     x
 })
@@ -226,7 +281,7 @@ gpsFromDb <- function(db, extraCols=NULL, bounds=NULL) {
     if(!('gpsData' %in% dbListTables(con))) {
         cat('No "gpsData" table found in database', basename(db), '\n')
         addChoice <- menu(title = 'Would you like to add gps data to this database?',
-             choices = c('Yes', 'No'))
+                          choices = c('Yes', 'No'))
         if(addChoice %in% c(0, 2)) {
             return(NULL)
         }