+++ NEWS.md
@@ -1,3 +1,19 @@
+## PAMpal 0.18.1
+
+- Updated to warnings in `addGps` to now be less overwhelming
+
+- Fixed bug for displaying some notes after combining with `bindStudies`
+
+- `addGps` now adds `gpsUncertainty` which says how many meters since to
+the closest GPS point. Useful to catch points where we have interpolated a lot.
+
+## PAMpal 0.18.0
+
+- Adding `addAnnotation` and related functions for annomate exporting
+
+- Explicitly marking joins that are expected to have multiple matches re:
+`dplyr::join`'s new behavior
+
 ## PAMpal 0.17.2
 
 - Smol change to not grab unnecessary columns in `addGps`